refactor
Remove cyclic dependencies between modules.
